Enable deferred and reuseport
This is only done for the first `listen`, because nginx will fail
otherwise. Idk why the first element is `2` instead of `1`, but as long
as it's only set on ONE server (doesn't have to be the first), it works.

Also, that `outer_loop` stuff is required because the nested loop will
override `loop`.

`deferred` enables `TCP_DEFER_ACCEPT` (see the corresponding man pages),
this basically allows for faster connection establishments.

`reuseport` enables `SO_REUSEPORT`, this has security implications
because another process on the machine can bind to port `80` as well,
allowing it to get some connections. However, when only nginx listens on
port `80`, every worker process can listen instead of just the master
process. This way, the kernel is able to distribute connections between
the worker processes in kernel space instead of the nginx master process
in user space, which should improve performance when many connections
are opened. More info here: https://lwn.net/Articles/542629/